LPS-192399 Bring SalesforceObjectEntryManagerImplTest back to object-storage-salesforce-test module
